TASK-43200 Update Drawer to use Common API
