+++ app-center-webapps/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -103,6 +103,9 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
       <module>eXoVueI18n</module>
     </depends>
     <depends>
+      <module>commonVueComponents</module>
+    </depends>
+    <depends>
       <module>jquery</module>
     </depends>
     <depends>
@@ -126,6 +129,9 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
       <module>eXoVueI18n</module>
     </depends>
     <depends>
+      <module>commonVueComponents</module>
+    </depends>
+    <depends>
       <module>jquery</module>
     </depends>
   </module>
@@ -144,6 +150,9 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
     </depends>
     <depends>
       <module>eXoVueI18n</module>
+    </depends>
+    <depends>
+      <module>commonVueComponents</module>
     </depends>
     <depends>
       <module>jquery</module>
@@ -172,6 +181,9 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
       <module>eXoVueI18n</module>
     </depends>
     <depends>
+      <module>commonVueComponents</module>
+    </depends>
+    <depends>
       <module>jquery</module>
     </depends>
   </module>
@@ -190,6 +202,9 @@ Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
     </depends>
     <depends>
       <module>eXoVueI18n</module>
+    </depends>
+    <depends>
+      <module>commonVueComponents</module>
     </depends>
     <depends>
       <module>jquery</module>
